prep clusters optim production rework. extract melo machina to mql module
